I am an idiot

+++ engine/default-shaders/vertex.glsl
@@ -8,7 +8,7 @@ in vec4 inMatProp2;
 in vec4 inMatProp3;
 in vec4 inMatProp4;
 in vec3 inNormal;
-in mediump float inMatIndex;
+in int inMatIndex;
 in vec2 inTexCoord;
 //attribute vec3 inNormal;
 
@@ -24,7 +24,7 @@ out vec4 matProp1;
 out vec4 matProp2;
 out vec4 matProp3;
 out vec4 matProp4;
-out float matIndex;
+out int matIndex;
 out vec3 normal;
 out vec3 position;
